Extract AST-parsing layer into parse-story-file.mjs
Move getNamedStringProperty, hasModifier, getDefaultExportObjectLiteral,
extractStoryExports, parseStoryFile, and their toKebabCase dependency into a
co-located scripts/parse-story-file.mjs module. This brings
generate-storybook-screenshots-config.mjs from 331 lines to 177 lines (under
the 200-line recommended max) and isolates the TypeScript AST-parsing concern
from CLI orchestration and story resolution.
